app/requirements.txt
Flask==3.0.3
instaloader==4.13.1
python-dotenv==1.0.1
